pyproject.toml
[tool.poetry]
name = "conduit"
version = "0.1.0"
description = "Pyramid and OpenAPI3 based RealWorld.io example."
authors = ["niteo.co <info@niteo.co>"]
license = "MIT"
classifiers = [
	"Programming Language :: Python",
	"Programming Language :: Python :: 3.8",
	"Framework :: Pyramid",
	"Topic :: Internet :: WWW/HTTP",
	"Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
	"License :: MIT",
]
homepage = "http://github.com/niteoweb/pyramid-realworld-example-app"
keywords = ["pyramid", "openapi", "realworld"]
packages = [
	{ include = "conduit", from = "src" },
]

[tool.poetry.plugins."paste.app_factory"]
main = "conduit:main"
test_suite = "conduit"

[tool.poetry.dependencies]
python = "~3.8"
argon2-cffi = "*"
gunicorn = "*"
mypy_extensions = "*"
passlib = "*"
psycopg2-binary = "*"
pyramid = "*"
pyramid_deferred_sqla = {git = "https://github.com/niteoweb/pyramid_deferred_sqla.git"}
pyramid_force_https = "*"
pyramid_heroku = "*"
pyramid_jwt = "*"
pyramid_openapi3 = "==0.5.3b2"
pyramid_redirect = "*"
python-slugify = "*"
raven = "*"
structlog = "*"
werkzeug = "*"  # remove after openapi-core 0.13.1
openapi-core = "==0.13.1"  # remove after pyramid_openapi3 0.5.0

[tool.poetry.dev-dependencies]
bandit = "*"
black = "==19.3b0"
codespell = "*"
coverage = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-deprecated = "*"
flake8-docstrings = "*"
flake8-isort = "*"
flake8-mock = "*"
flake8-mutable = "*"
flake8-plone-hasattr = "*"
flake8-print = "*"
flake8-self = "*"
flake8-super-call = "*"
flake8-tuple = "*"
freezegun = "*"
mypy = "*"
pre-commit = "*"
pre-commit-hooks = "*"
pytest = "*"
pytest-bdd = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-randomly = "*"
pytest-testmon = "*"
pytest-watch = "*"
requests-html = "*"
responses = "*"
testfixtures = "*"
typecov = "*"
waitress = "*"
webtest = "*"
npm = "^0.1.1"
sqlalchemy-stubs = "^0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
